+++ R/rowIQRs.R
@@ -32,7 +32,7 @@ rowIQRs <- function(x, rows = NULL, cols = NULL, na.rm = FALSE, ..., useNames = 
 
   # Remove attributes
   if (is.na(useNames)) {
-    deprecatedUseNamesNA()
+    defunctUseNamesNA()
     attributes(ans) <- NULL
   }
   
@@ -50,7 +50,7 @@ colIQRs <- function(x, rows = NULL, cols = NULL, na.rm = FALSE, ..., useNames = 
 
   # Remove attributes
   if (is.na(useNames)) {
-    deprecatedUseNamesNA()
+    defunctUseNamesNA()
     attributes(ans) <- NULL
   }