+++ R/rowMedians.R
@@ -41,7 +41,7 @@
 #' @export
 rowMedians <- function(x, rows = NULL, cols = NULL, na.rm = FALSE,
                        dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   has_nas <- TRUE  # Add as an argument? /2007-08-24
   
   .Call(C_rowMedians, x, dim., rows, cols, na.rm, has_nas, TRUE, useNames)
@@ -51,7 +51,7 @@ rowMedians <- function(x, rows = NULL, cols = NULL, na.rm = FALSE,
 #' @export
 colMedians <- function(x, rows = NULL, cols = NULL, na.rm = FALSE,
                        dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   has_nas <- TRUE  # Add as an argument? /2007-08-24
   
   .Call(C_rowMedians, x, dim., rows, cols, na.rm, has_nas, FALSE, useNames)