+++ R/rowOrderStats.R
@@ -36,7 +36,7 @@ rowOrderStats <- function(x, rows = NULL, cols = NULL, which,
     stop(sprintf("Argument '%s' must not contain missing values", "x"))
   }
 
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_rowOrderStats, x, dim., rows, cols, which, useNames)
 }
 
@@ -51,6 +51,6 @@ colOrderStats <- function(x, rows = NULL, cols = NULL, which,
     stop(sprintf("Argument '%s' must not contain missing values", "x"))
   }
 
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_colOrderStats, x, dim., rows, cols, which, useNames)
 }