+++ R/rowRanks.R
@@ -103,7 +103,7 @@ rowRanks <- function(x, rows = NULL, cols = NULL,
     stop(sprintf("Unknown value of argument '%s': %s", "ties.method", ties.method))
   }
 
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   
   # byrow = TRUE
   .Call(C_rowRanksWithTies, x, dim., rows, cols, ties_method, TRUE, useNames)
@@ -134,7 +134,7 @@ colRanks <- function(x, rows = NULL, cols = NULL,
     stop(sprintf("Unknown value of argument '%s': %s", "ties.method", ties.method))
   }
 
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   
   # byrow = FALSE
   y <- .Call(C_rowRanksWithTies, x, dim., rows, cols, ties_method, FALSE, useNames)