+++ R/rowWeightedMedians.R
@@ -48,7 +48,7 @@ rowWeightedMedians <- function(x, w = NULL, rows = NULL, cols = NULL,
     }
   }
 
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
 
   # Apply subset on x
   if (!is.null(rows) && !is.null(cols)) x <- x[rows, cols, drop = FALSE]
@@ -108,7 +108,7 @@ colWeightedMedians <- function(x, w = NULL, rows = NULL, cols = NULL,
     }
   }
 
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
 
   # Apply subset on x
   if (!is.null(rows) && !is.null(cols)) x <- x[rows, cols, drop = FALSE]